requirements.txt
# Spécifique à dbt
dbt-core >= 1.5.0
dbt-snowflake >= 1.5.0

# Nécessaire pour streamlit
matplotlib==3.7.2
openpyxl==3.1.2
pandas==2.0.3
plotly==5.15.0
pyarrow==10.0.1
snowflake-connector-python==3.0.4
snowflake-snowpark-python==1.5.1
streamlit>=1.24.0